requirements.txt
openai==0.28.1
SpeechRecognition==3.10.4
pyaudio==0.2.14
gTTS==2.5.1
pygame==2.5.2
python-dotenv==1.0.1